more

+++ packages/protocol/test/layer2/based/TaikoAnchor.t.sol
@@ -32,7 +32,7 @@ contract TestTaikoAnchor is Layer2Test {
                     address(signalService),
                     0,
                     0,
-                    uint48(100), // maxCheckpointStackSize
+                    uint16(100), // maxCheckpointHistory
                     address(0) // bondManager
                 )
             ),
